(pC-3222): clean mediation table and mediation model
